src/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone
langchain-openai
langchain-pinecone
supabase
requests
twilio
python-multipart